backend/requirements.txt
livekit-agents>=0.9.0
livekit-plugins-deepgram>=0.9.0
livekit-plugins-cartesia>=0.9.0
livekit-plugins-openai>=0.9.0
livekit>=0.10.0
openai>=1.12.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
flask>=3.0.0
flask-cors>=4.0.0
aiohttp>=3.9.0
